refactor: implement useCouponManagement hook and restructure OrderPage for improved coupon handling
